United back K8sYamlParser with K8sPolicyYamlParser. Fixed syntax warnings. Moved the check of the target_pods' named ports to a later stage (after populating target_pods).
